requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.1
aiohttp>=3.9.0
pytz>=2024.1
icalendar>=5.0.11
